pyproject.toml
[project]
name = "keywordlens-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.128.0",
    "lxml>=6.0.2",
    "numpy>=2.4.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "selenium>=4.39.0",
    "uvicorn>=0.40.0",
    "webdriver-manager>=4.0.2",
]
